feat(2024): add CPU and Memory usage profiling
