+++ 2024/Day09/main.py
@@ -6,7 +6,7 @@ sys.path.append(".")
 from solutions import part1, part2
 
 from utils.inputs import get_inputs
-from utils.timings import profile_run
+from utils.profiling import profile_run
 
 if __name__ == "__main__":
     input_path = f"{os.path.dirname(os.path.realpath(__file__))}/input.txt"
